0.0.2 Rename to Okolni Source Query (instead of Okolni Source Engine Query)
