+++ Okolni.Source.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29123.88
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Okolni.SourceEngine.Query", "Okolni.SourceEngine.Query\Okolni.SourceEngine.Query.csproj", "{EA8C8BB3-74BF-4C47-B559-EE39631B2CD0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Okolni.Source.Query", "Okolni.Source.Query\Okolni.Source.Query.csproj", "{EA8C8BB3-74BF-4C47-B559-EE39631B2CD0}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Okolni.SourceEngine.Example", "Okolni.SourceEngine.Example\Okolni.SourceEngine.Example.csproj", "{82486710-5210-40C0-AA00-2C425735FDA6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Okolni.Source.Example", "Okolni.Source.Example\Okolni.Source.Example.csproj", "{82486710-5210-40C0-AA00-2C425735FDA6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
